Split case triggers into subpackage by use case
