+++ test/core/use_cases/triggers/case/__init__.py
@@ -1,0 +1,1 @@
+"""Case trigger use-case tests."""
