+++ test/core/use_cases/triggers/case/test_add_object.py
