+++ test/core/use_cases/triggers/case/test_create.py
