+++ vultron/core/use_cases/triggers/case/__init__.py
@@ -1,0 +1,53 @@
+#!/usr/bin/env python
+
+#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
+#  - see Contributors.md for a full list of Contributors
+#  - see ContributionInstructions.md for information on how you can Contribute to this project
+#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
+#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
+#  with this Software or contact permission@sei.cmu.edu for full terms.
+#  Created, in part, with funding and support from the United States Government
+#  (see Acknowledgments file). This program may include and/or can make use of
+#  certain third party source code, object code, documentation and other files
+#  ("Third Party Software"). See LICENSE.md for more details.
+#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
+#  U.S. Patent and Trademark Office by Carnegie Mellon University
+
+"""Case trigger use-case package.
+
+This package replaces the former ``triggers/case.py`` flat module with
+one-module-per-use-case while preserving existing imports such as:
+
+``from vultron.core.use_cases.triggers.case import SvcEngageCaseUseCase``.
+"""
+
+from vultron.core.use_cases.triggers.requests import (
+    AddObjectToCaseTriggerRequest,
+    AddParticipantStatusTriggerRequest,
+    AddReportToCaseTriggerRequest,
+    CreateCaseTriggerRequest,
+    DeferCaseTriggerRequest,
+    EngageCaseTriggerRequest,
+)
+
+from .add_object import SvcAddObjectToCaseUseCase
+from .add_participant_status import SvcAddParticipantStatusUseCase
+from .add_report import SvcAddReportToCaseUseCase
+from .create import SvcCreateCaseUseCase
+from .defer import SvcDeferCaseUseCase
+from .engage import SvcEngageCaseUseCase
+
+__all__ = [
+    "AddObjectToCaseTriggerRequest",
+    "AddParticipantStatusTriggerRequest",
+    "AddReportToCaseTriggerRequest",
+    "CreateCaseTriggerRequest",
+    "DeferCaseTriggerRequest",
+    "EngageCaseTriggerRequest",
+    "SvcAddObjectToCaseUseCase",
+    "SvcAddParticipantStatusUseCase",
+    "SvcAddReportToCaseUseCase",
+    "SvcCreateCaseUseCase",
+    "SvcDeferCaseUseCase",
+    "SvcEngageCaseUseCase",
+]
