+++ vultron/core/use_cases/triggers/case/add_object.py
@@ -1,0 +1,94 @@
+#!/usr/bin/env python
+
+#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
+#  - see Contributors.md for a full list of Contributors
+#  - see ContributionInstructions.md for information on how you can Contribute to this project
+#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
+#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
+#  with this Software or contact permission@sei.cmu.edu for full terms.
+#  Created, in part, with funding and support from the United States Government
+#  (see Acknowledgments file). This program may include and/or can make use of
+#  certain third party source code, object code, documentation and other files
+#  ("Third Party Software"). See LICENSE.md for more details.
+#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
+#  U.S. Patent and Trademark Office by Carnegie Mellon University
+
+import logging
+from typing import TYPE_CHECKING, Any
+
+from vultron.core.ports.case_persistence import CaseOutboxPersistence
+from vultron.core.use_cases.triggers._helpers import (
+    add_activity_to_outbox,
+    resolve_actor,
+    resolve_case,
+)
+from vultron.core.use_cases.triggers.requests import (
+    AddObjectToCaseTriggerRequest,
+)
+from vultron.errors import VultronNotFoundError
+
+if TYPE_CHECKING:
+    from vultron.core.ports.trigger_activity import TriggerActivityPort
+
+logger = logging.getLogger(__name__)
+
+
+class SvcAddObjectToCaseUseCase:
+    """Add any existing AS2 object to a case (general-purpose).
+
+    Reads the object by ``object_id`` from the datalayer, creates a generic
+    ``Add(object, target=case)`` activity, and queues it in the actor's
+    outbox. The object must already exist; this use case does not create it.
+
+    Type-specific wrappers (e.g., ``SvcAddReportToCaseUseCase``) delegate
+    here after performing their own validation (TRIG-10-002).
+
+    Implements: TRIG-10-001.
+    """
+
+    def __init__(
+        self,
+        dl: CaseOutboxPersistence,
+        request: AddObjectToCaseTriggerRequest,
+        trigger_activity: "TriggerActivityPort | None" = None,
+    ) -> None:
+        self._dl = dl
+        self._request = request
+        self._trigger_activity = trigger_activity
+
+    def execute(self) -> dict[str, Any]:
+        actor_id = self._request.actor_id
+        case_id = self._request.case_id
+        object_id = self._request.object_id
+        dl = self._dl
+
+        actor = resolve_actor(actor_id, dl)
+        actor_id = actor.id_
+
+        resolve_case(case_id, dl)
+
+        raw = dl.read(object_id)
+        if raw is None:
+            raise VultronNotFoundError("AS2Object", object_id)
+
+        if self._trigger_activity is None:
+            raise RuntimeError(
+                "SvcAddObjectToCaseUseCase requires a TriggerActivityPort"
+            )
+
+        activity_id, activity_dict = self._trigger_activity.add_object_to_case(
+            actor=actor_id,
+            object_id=object_id,
+            case_id=case_id,
+        )
+
+        add_activity_to_outbox(actor_id, activity_id, dl)
+
+        logger.info(
+            "Actor '%s' added object '%s' to case '%s'",
+            actor_id,
+            object_id,
+            case_id,
+        )
+
+        return {"activity": activity_dict}
