+++ vultron/core/use_cases/triggers/case/create.py
@@ -1,0 +1,99 @@
+#!/usr/bin/env python
+
+#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
+#  - see Contributors.md for a full list of Contributors
+#  - see ContributionInstructions.md for information on how you can Contribute to this project
+#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
+#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
+#  with this Software or contact permission@sei.cmu.edu for full terms.
+#  Created, in part, with funding and support from the United States Government
+#  (see Acknowledgments file). This program may include and/or can make use of
+#  certain third party source code, object code, documentation and other files
+#  ("Third Party Software"). See LICENSE.md for more details.
+#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
+#  U.S. Patent and Trademark Office by Carnegie Mellon University
+
+import logging
+from typing import TYPE_CHECKING, Any
+
+from vultron.core.models.case import VultronCase
+from vultron.core.ports.case_persistence import CaseOutboxPersistence
+from vultron.core.use_cases.triggers._helpers import (
+    add_activity_to_outbox,
+    resolve_actor,
+)
+from vultron.core.use_cases.triggers.requests import (
+    CreateCaseTriggerRequest,
+)
+from vultron.errors import VultronNotFoundError, VultronValidationError
+
+if TYPE_CHECKING:
+    from vultron.core.ports.trigger_activity import TriggerActivityPort
+
+logger = logging.getLogger(__name__)
+
+
+class SvcCreateCaseUseCase:
+    """Create a new VulnerabilityCase and emit a CreateCaseActivity.
+
+    The actor creates a local case and queues the activity for delivery to
+    the CaseActor inbox. An optional report_id links an existing
+    VulnerabilityReport to the new case.
+    """
+
+    def __init__(
+        self,
+        dl: CaseOutboxPersistence,
+        request: CreateCaseTriggerRequest,
+        trigger_activity: "TriggerActivityPort | None" = None,
+    ) -> None:
+        self._dl = dl
+        self._request = request
+        self._trigger_activity = trigger_activity
+
+    def execute(self) -> dict[str, Any]:
+        actor_id = self._request.actor_id
+        actor = resolve_actor(actor_id, self._dl)
+        actor_id = actor.id_
+
+        case = VultronCase(
+            name=self._request.name,
+            content=self._request.content,
+            attributed_to=actor.id_,
+        )
+
+        if self._request.report_id is not None:
+            raw = self._dl.read(self._request.report_id)
+            if raw is None:
+                raise VultronNotFoundError(
+                    "VulnerabilityReport", self._request.report_id
+                )
+            if getattr(raw, "type_", "") != "VulnerabilityReport":
+                raise VultronValidationError(
+                    f"'{self._request.report_id}' is not a VulnerabilityReport"
+                )
+            raw_id = getattr(raw, "id_", None) or self._request.report_id
+            case.vulnerability_reports.append(raw_id)
+
+        self._dl.create(case)
+
+        if self._trigger_activity is None:
+            raise RuntimeError(
+                "SvcCreateCaseUseCase requires a TriggerActivityPort"
+            )
+
+        activity_id, activity_dict = self._trigger_activity.create_case(
+            case_id=case.id_,
+            actor=actor.id_,
+        )
+
+        add_activity_to_outbox(actor_id, activity_id, self._dl)
+
+        logger.info(
+            "Actor '%s' created case '%s' (CreateCaseActivity '%s')",
+            actor_id,
+            case.id_,
+            activity_id,
+        )
+
+        return {"activity": activity_dict}
